requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0

# AI・機械学習関連
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
sentence-transformers==2.2.2
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3

# 画像処理・認識関連
ultralytics==8.0.196
easyocr==1.7.0

# ベクトルデータベース関連
pgvector==0.2.3

# その他
python-multipart==0.0.6
bcrypt==4.0.1
PyJWT==2.8.0
python-magic==0.4.27

# テスト関連
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# セキュリティ・品質管理
bandit==1.7.5
safety>=2.4.0

# コード品質
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
